gvae run

+++ ptc_mr_prev/deepinfomax_gvae_nonodedata.py
@@ -177,7 +177,7 @@ class GcnInfomax(nn.Module):
                 data.to(device)
                 x, edge_index, batch = data.x, data.edge_index, data.batch
                 if not dataset.num_features:
-                    x = torch.ones((batch.shape[0],5)).double().to(device)
+                    x = torch.ones((batch.shape[0],5)).to(device)
                 node_mu, node_logvar = self.encoder(x, edge_index, batch)
 
                 #node_latent_embeddings = reparameterize(training=False, mu=node_mu, logvar=node_logvar)
@@ -303,7 +303,7 @@ if __name__ == '__main__':
                 data = data.to(device)
                 if not dataset.num_features:
 
-                    data.x = torch.ones((data.batch.shape[0], 5)).double().to(device)
+                    data.x = torch.ones((data.batch.shape[0], 5)).to(device)
                 optimizer.zero_grad()
                 recon_loss, kl_class, kl_node = model(data.x, data.edge_index, data.batch, data.num_graphs)
                 recon_loss_all += recon_loss
